Add AppInit container as our 'auth' stage
This initializes the app and routes to the right screen based
on whether an identity exists, or registration is already in progress

similar to https://reactnavigation.org/docs/en/auth-flow.html

Also get rid of all that back button handling logic, the defaults
already work great if the routes are set up right and app init logic
happens at the right point (under react-navigation, as a separate route)
