+++ src/ui/generic/index.ts
@@ -1,5 +1,4 @@
 export { Exception } from './exception'
-export { LoadingScreen } from './loading'
 export { LoadingSpinner } from './loadingSpinner'
 
 export const SCROLL_PADDING_BOTTOM = 36
